Working implementation of @Autowired
As Sim is implemented by to classes Airtel and Jio, Spring is not able
to distinguish which bean to inject, hence by declaring one bean with
@primary or giving @Qualifier("nameInCamelCase") during autowiring
spring knows which one to inject

Note: When we run the above code, Spring will throw a NoUniqueBeanDefinitionException because it finds two beans (Airtel and Jio) of type Sim. To resolve this, we need to specify which bean to inject.